+++ CHANGELOG.md
@@ -6,6 +6,21 @@ All rights reserved.
 
 *THIS PACKAGE IS **PRE-RELEASE**, IN ACTIVE DEVELOPMENT AND SUBJECT TO DAILY BREAKING CHANGES.*
 
+## 0.0.7
+**Beta**
+
+### *Breaking changes*
+* New method `TermProperties.lemmasOf`.
+* New method `TermProperties.allLemmas`.
+* New method `TermProperties.lemmasMap`.
+* New method `Thesaurus.lemmasOf`.
+
+### *Updated*
+* Documentation.
+* Tests.
+* Examples.
+* Dependencies.
+
 ## 0.0.6
 **Beta**